Split to separate declarations into header
